// src/types/Toolbar.ts
export interface Toolbar {
	getDropIndex: ((x, y) => any)
	getIconSize: (() => any)
	getItemIndex: ((item) => any)
	getNItems: (() => any)
	getNthItem: ((n) => any)
	getReliefStyle: (() => any)
	getShowArrow: (() => any)
	getStyle: (() => any)
	insert: ((item, pos) => any)
	setDropHighlightItem: ((tool_item, index_) => any)
	setIconSize: ((icon_size) => any)
	setShowArrow: ((show_arrow) => any)
	setStyle: ((style) => any)
	unsetIconSize: (() => any)
	unsetStyle: (() => any)
	add: ((item) => any)
	show: (() => any)
	hide: (() => any)
}