// src/types/Label.ts
export interface Label {
	show: (() => any)
	hide: (() => any)
	getAngle: (() => any)
	getAttributes: (() => any)
	getCurrentUri: (() => any)
	getEllipsize: (() => any)
	getJustify: (() => any)
	getLabel: (() => any)
	getLayout: (() => any)
	getLayoutOffsets: (() => any)
	getLineWrap: (() => any)
	getLineWrapMode: (() => any)
	getLines: (() => any)
	getMaxWidthChars: (() => any)
	getMnemonicKeyval: (() => any)
	getMnemonicWidget: (() => any)
	getSelectable: (() => any)
	getSelectionBounds: (() => any)
	getSingleLineMode: (() => any)
	getText: (() => any)
	getTrackVisitedLinks: (() => any)
	getUseMarkup: (() => any)
	getUseUnderline: (() => any)
	getWidthChars: (() => any)
	getXalign: (() => any)
	getYalign: (() => any)
	selectRegion: ((start_offset, end_offset) => any)
	setAngle: ((angle) => any)
	setAttributes: ((attrs) => any)
	setEllipsize: ((mode) => any)
	setJustify: ((jtype) => any)
	setLabel: ((str) => any)
	setLineWrap: ((wrap) => any)
	setLineWrapMode: ((wrap_mode) => any)
	setLines: ((lines) => any)
	setMarkup: ((str) => any)
	setMarkupWithMnemonic: ((str) => any)
	setMaxWidthChars: ((n_chars) => any)
	setMnemonicWidget: ((widget) => any)
	setPattern: ((pattern) => any)
	setSelectable: ((setting) => any)
	setSingleLineMode: ((single_line_mode) => any)
	setText: ((str) => any)
	setTextWithMnemonic: ((str) => any)
	setTrackVisitedLinks: ((track_links) => any)
	setUseMarkup: ((setting) => any)
	setUseUnderline: ((setting) => any)
	setWidthChars: ((n_chars) => any)
	setXalign: ((xalign) => any)
	setYalign: ((yalign) => any)

	label: string
}