// src/types/ComboBoxText.ts
export interface ComboBoxText {
	show: (() => any)
	hide: (() => any)
	getActivatesDefault: (() => any)
	getAlignment: (() => any)
	getAttributes: (() => any)
	getBuffer: (() => any)
	getCompletion: (() => any)
	getCurrentIconDragSource: (() => any)
	getCursorHadjustment: (() => any)
	getHasFrame: (() => any)
	getIconActivatable: ((icon_pos) => any)
	getIconArea: ((icon_pos) => any)
	getIconAtPos: ((x, y) => any)
	getIconGicon: ((icon_pos) => any)
	getIconName: ((icon_pos) => any)
	getIconPixbuf: ((icon_pos) => any)
	getIconSensitive: ((icon_pos) => any)
	getIconStock: ((icon_pos) => any)
	getIconStorageType: ((icon_pos) => any)
	getIconTooltipMarkup: ((icon_pos) => any)
	getIconTooltipText: ((icon_pos) => any)
	getInnerBorder: (() => any)
	getInputHints: (() => any)
	getInputPurpose: (() => any)
	getInvisibleChar: (() => any)
	getLayout: (() => any)
	getLayoutOffsets: (() => any)
	getMaxLength: (() => any)
	getMaxWidthChars: (() => any)
	getOverwriteMode: (() => any)
	getPlaceholderText: (() => any)
	getProgressFraction: (() => any)
	getProgressPulseStep: (() => any)
	getTabs: (() => any)
	getText: (() => any)
	getTextArea: (() => any)
	getTextLength: (() => any)
	getVisibility: (() => any)
	getWidthChars: (() => any)
	grabFocusWithoutSelecting: (() => any)
	imContextFilterKeypress: ((event) => any)
	layoutIndexToTextIndex: ((layout_index) => any)
	progressPulse: (() => any)
	resetImContext: (() => any)
	setActivatesDefault: ((setting) => any)
	setAlignment: ((xalign) => any)
	setAttributes: ((attrs) => any)
	setBuffer: ((buffer) => any)
	setCompletion: ((completion) => any)
	setCursorHadjustment: ((adjustment) => any)
	setHasFrame: ((setting) => any)
	setIconActivatable: ((icon_pos, activatable) => any)
	setIconDragSource: ((icon_pos, target_list, actions) => any)
	setIconFromGicon: ((icon_pos, icon) => any)
	setIconFromIconName: ((icon_pos, icon_name) => any)
	setIconFromPixbuf: ((icon_pos, pixbuf) => any)
	setIconFromStock: ((icon_pos, stock_id) => any)
	setIconSensitive: ((icon_pos, sensitive) => any)
	setIconTooltipMarkup: ((icon_pos, tooltip) => any)
	setIconTooltipText: ((icon_pos, tooltip) => any)
	setInnerBorder: ((border) => any)
	setInputHints: ((hints) => any)
	setInputPurpose: ((purpose) => any)
	setInvisibleChar: ((ch) => any)
	setMaxLength: ((max) => any)
	setMaxWidthChars: ((n_chars) => any)
	setOverwriteMode: ((overwrite) => any)
	setPlaceholderText: ((text) => any)
	setProgressFraction: ((fraction) => any)
	setProgressPulseStep: ((fraction) => any)
	setTabs: ((tabs) => any)
	setText: ((text) => any)
	setVisibility: ((visible) => any)
	setWidthChars: ((n_chars) => any)
	textIndexToLayoutIndex: ((text_index) => any)
	unsetInvisibleChar: (() => any)
	appendText: ((text: string) => any)
	on: ((event: string, fallback: any) => any)
	getActiveText: (() => string)
	removeAll: (() => any)
}