// src/types/WebView.ts
export interface WebView {
	show: (() => any)
	hide: (() => any)
	canExecuteEditingCommand: ((command, cancellable, callback, user_data) => any)
	canExecuteEditingCommandFinish: ((result) => any)
	canGoBack: (() => any)
	canGoForward: (() => any)
	canShowMimeType: ((mime_type) => any)
	downloadUri: ((uri) => any)
	executeEditingCommand: ((command) => any)
	executeEditingCommandWithArgument: ((command, argument) => any)
	getAutomationPresentationType: (() => any)
	getBackForwardList: (() => any)
	getBackgroundColor: (() => any)
	getContext: (() => any)
	getCustomCharset: (() => any)
	getEditorState: (() => any)
	getEstimatedLoadProgress: (() => any)
	getFavicon: (() => any)
	getFindController: (() => any)
	getInputMethodContext: (() => any)
	getInspector: (() => any)
	getIsMuted: (() => any)
	getMainResource: (() => any)
	getPageId: (() => any)
	getSessionState: (() => any)
	getSettings: (() => any)
	getSnapshot: ((region, options, cancellable, callback, user_data) => any)
	getSnapshotFinish: ((result) => any)
	getTitle: (() => any)
	getTlsInfo: (() => any)
	getUri: (() => any)
	getUserContentManager: (() => any)
	getWebsiteDataManager: (() => any)
	getWebsitePolicies: (() => any)
	getWindowProperties: (() => any)
	getZoomLevel: (() => any)
	goBack: (() => any)
	goForward: (() => any)
	goToBackForwardListItem: ((list_item) => any)
	isControlledByAutomation: (() => any)
	isEditable: (() => any)
	isEphemeral: (() => any)
	isLoading: (() => any)
	isPlayingAudio: (() => any)
	loadAlternateHtml: ((content, content_uri, base_uri) => any)
	loadBytes: ((bytes, mime_type, encoding, base_uri) => any)
	loadHtml: ((content, base_uri) => any)
	loadPlainText: ((plain_text) => any)
	loadRequest: ((request) => any)
	loadUri: ((uri) => any)
	reload: (() => any)
	reloadBypassCache: (() => any)
	restoreSessionState: ((state) => any)
	runJavascript: ((script, cancellable, callback, user_data) => any)
	runJavascriptFinish: ((result) => any)
	runJavascriptFromGresource: ((resource, cancellable, callback, user_data) => any)
	runJavascriptFromGresourceFinish: ((result) => any)
	runJavascriptInWorld: ((script, world_name, cancellable, callback, user_data) => any)
	runJavascriptInWorldFinish: ((result) => any)
	save: ((save_mode, cancellable, callback, user_data) => any)
	saveFinish: ((result) => any)
	saveToFile: ((file, save_mode, cancellable, callback, user_data) => any)
	saveToFileFinish: ((result) => any)
	sendMessageToPage: ((message, cancellable, callback, user_data) => any)
	sendMessageToPageFinish: ((result) => any)
	setBackgroundColor: ((rgba) => any)
	setCustomCharset: ((charset) => any)
	setEditable: ((editable) => any)
	setInputMethodContext: ((context) => any)
	setIsMuted: ((muted) => any)
	setSettings: ((settings) => any)
	setZoomLevel: ((zoom_level) => any)
	stopLoading: (() => any)
	tryClose: (() => any)
	on: ((event: string, fallback: any) => any)
}